# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CLinic Software Api ",
        "description": "API Description",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/reservations/check-conflict": {
            "post": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Check for reservation conflicts",
                "operationId": "906d0c3c44c4f3a7bba61f459837ac20",
                "requestBody": {
                    "description": "Data to check for conflicts",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CheckReservationConflictRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Conflict check result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "conflict_exists": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reservations": {
            "post": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Store a newly created reservation",
                "operationId": "01e3d21dc330a31de025203bc142ffa3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReservationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Resource created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReservationResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reservations/{reservation}": {
            "get": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Display the specified reservation",
                "operationId": "2aefc564f94e483fb93920c2935eeafe",
                "parameters": [
                    {
                        "name": "reservation",
                        "in": "path",
                        "description": "Reservation ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReservationResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Update the specified reservation",
                "operationId": "459cb3958996a0266b294be06a643db0",
                "parameters": [
                    {
                        "name": "reservation",
                        "in": "path",
                        "description": "Reservation ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReservationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Resource updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReservationResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Remove the specified reservation",
                "operationId": "9cb75e88bfb3a102881ea262dde33019",
                "parameters": [
                    {
                        "name": "reservation",
                        "in": "path",
                        "description": "Reservation ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resource deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReservationResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Reservation not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/treatments/{treatment}": {
            "put": {
                "tags": [
                    "Treatments"
                ],
                "summary": "Update a treatment",
                "operationId": "0c60412a22ee00cd2b048e5cfc88bbc8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Treatment Name"
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "#000000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Treatment updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TreatmentResource"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Treatments"
                ],
                "summary": "Delete a treatment",
                "operationId": "092366bdd563ae92463e16887f0465d1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Treatment deleted"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "CheckReservationConflictRequest": {
                "required": [
                    "start",
                    "end"
                ],
                "properties": {
                    "clinicId": {
                        "description": "The ID of the clinic",
                        "type": "string",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "start": {
                        "description": "The start time of the reservation",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-01T09:00:00Z"
                    },
                    "end": {
                        "description": "The end time of the reservation",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-01T10:00:00Z"
                    },
                    "reservationId": {
                        "description": "The ID of the reservation (nullable for new reservations)",
                        "type": "string",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    }
                },
                "type": "object"
            },
...
            "ReservationRequest": {
                "description": "Request data for creating or updating a reservation",
                "required": [
                    "start",
                    "end",
                    "type",
                    "patientId"
                ],
                "properties": {
                    "start": {
                        "description": "Reservation start time",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-12-12T02:30:00Z"
                    },
                    "end": {
                        "description": "Reservation end time",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-12-12T03:30:00Z"
                    },
                    "patientId": {
                        "description": "Patient UUID",
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "type": {
                        "description": "Reservation type",
                        "type": "string",
                        "enum": [
                            "surgery",
                            "appointment",
                            "inspection"
                        ],
                        "example": "surgery"
                    },
                    "status": {
                        "description": "Reservation status",
                        "type": "string",
                        "default": "income",
                        "enum": [
                            "income",
                            "check",
                            "dismiss",
                            "cancelled"
                        ],
                        "example": "income"
                    },
                    "doctorId": {
                        "description": "Doctor UUID",
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440001"
                    },
                    "medicalCaseId": {
                        "description": "Medical Case UUID",
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440002"
                    },
                    "clinicId": {
                        "description": "Clinic UUID",
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440003"
                    }
                },
                "type": "object"
            },
...
            "DoctorResource": {
                "title": "Doctor Resource",
                "description": "Doctor resource representation",
                "properties": {
                    "string": {
                        "description": "The unique identifier of the doctor",
                        "type": "string"
                    },
                    "fullName": {
                        "description": "The full name of the doctor",
                        "type": "string"
                    },
                    "avatar": {
                        "$ref": "#/components/schemas/MediaResource"
                    }
                },
                "type": "object"
            },
...
            "MedicalCaseResource": {
                "title": "Medical Case Resource",
                "description": "Medical Case Resource",
                "properties": {
                    "id": {
                        "description": "The ID of the medical case",
                        "type": "integer"
                    },
                    "name": {
                        "description": "The name of the medical case",
                        "type": "string"
                    },
                    "description": {
                        "description": "The description of the medical case",
                        "type": "string"
                    },
                    "total": {
                        "description": "The total amount associated with the medical case",
                        "type": "number",
                        "format": "float"
                    },
                    "date": {
                        "description": "The date of the medical case",
                        "type": "string",
                        "format": "date-time"
                    },
                    "createdBy": {
                        "$ref": "#/components/schemas/UserResource"
                    }
                },
                "type": "object"
            },
...
            "PatientResource": {
                "title": "PatientResource",
                "description": "Patient resource representation",
                "properties": {
                    "id": {
                        "description": "Unique identifier for the patient",
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "firstName": {
                        "description": "First name of the patient",
                        "type": "string",
                        "example": "John"
                    },
                    "lastName": {
                        "description": "Last name of the patient",
                        "type": "string",
                        "example": "Doe"
                    },
                    "avatar": {
                        "$ref": "#/components/schemas/MediaResource"
                    },
                    "age": {
                        "description": "Age of the patient",
                        "type": "integer",
                        "example": 30
                    },
                    "fatherName": {
                        "description": "Father's name of the patient",
                        "type": "string",
                        "example": "Robert"
                    },
                    "motherName": {
                        "description": "Mother's name of the patient",
                        "type": "string",
                        "example": "Jane"
                    },
                    "birth": {
                        "description": "Birth date of the patient",
                        "type": "string",
                        "format": "date",
                        "example": "1993-05-15"
                    },
                    "phone": {
                        "description": "Phone number of the patient",
                        "type": "string",
                        "example": "+1234567890"
                    },
                    "gender": {
                        "description": "Gender of the patient",
                        "type": "string",
                        "enum": [
                            "male",
                            "female"
                        ],
                        "example": "male"
                    },
                    "notes": {
                        "description": "Additional notes about the patient",
                        "type": "string",
                        "example": "Patient has allergies"
                    },
                    "nationalNumber": {
                        "description": "National identification number of the patient",
                        "type": "string",
                        "example": "123456789"
                    },
                    "address": {
                        "description": "Address of the patient",
                        "type": "string",
                        "example": "123 Main St, Anytown, USA"
                    },
                    "createdAt": {
                        "description": "Creation timestamp of the patient record",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-12T07:20:50.52Z"
                    },
                    "nextReservation": {
                        "description": "Next reservation date for the patient",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-11-15T10:00:00.00Z"
                    },
                    "lastReservation": {
                        "description": "Last reservation date for the patient",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-09-10T14:30:00.00Z"
                    },
                    "permanentIlls": {
                        "description": "List of permanent illnesses associated with the patient",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/IllResource"
                        }
                    },
                    "permanentMedicines": {
                        "description": "List of permanent medicines associated with the patient",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MedicineResource"
                        }
                    },
                    "media": {
                        "description": "List of media files associated with the patient (excluding avatar)",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MediaResource"
                        }
                    }
                },
                "type": "object"
            },
...
            "ReservationResource": {
                "title": "Reservation Resource",
                "description": "Reservation resource representation",
                "properties": {
                    "id": {
                        "description": "Reservation ID",
                        "type": "string"
                    },
                    "start": {
                        "description": "Start date and time of the reservation",
                        "type": "string",
                        "format": "date-time"
                    },
                    "end": {
                        "description": "End date and time of the reservation",
                        "type": "string",
                        "format": "date-time"
                    },
                    "type": {
                        "description": "Type of the reservation",
                        "type": "string"
                    },
                    "status": {
                        "description": "Status of the reservation",
                        "type": "string"
                    },
                    "patient": {
                        "$ref": "#/components/schemas/PatientResource"
                    },
                    "doctor": {
                        "$ref": "#/components/schemas/DoctorResource"
                    },
                    "medicalCase": {
                        "$ref": "#/components/schemas/MedicalCaseResource"
                    },
                    "createdAt": {
                        "description": "Creation timestamp of the reservation",
                        "type": "string",
                        "format": "date-time"
                    },
                    "pastReservationsCount": {
                        "description": "Count of past reservations",
                        "type": "integer"
                    },
                    "upComingReservationCount": {
                        "description": "Count of upcoming reservations",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "TreatmentResource": {
                "title": "Treatment Resource",
                "description": "Treatment resource representation",
                "properties": {
                    "id": {
                        "description": "The unique identifier of the treatment",
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "description": "The name of the treatment",
                        "type": "string"
                    },
                    "color": {
                        "description": "The color associated with the treatment",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}